@@ -17,7 +17,7 @@ extension BlocModuleExtension on Module {
   @deprecated
   ProviderDefinition<T> bloc<T extends Disposable>(
     ProviderCreate<T> definition, {
-    Qualifier qualifier,
+    Qualifier? qualifier,
     bool createdAtStart = false,
     bool override = false,
   }) {
@@ -26,7 +26,8 @@ extension BlocModuleExtension on Module {
         createdAtStart: createdAtStart,
         override: override);
 
-    beanDefinition.onClose((bloc) => bloc.dispose());
+    beanDefinition.onClose((bloc) => bloc.dispose(),
+        onDisposeUnitialized: () {});
     return beanDefinition;
   }
 
@@ -42,7 +43,7 @@ extension BlocModuleExtension on Module {
   @deprecated
   ProviderDefinition<T> scopeOneBloc<T extends Disposable, TScope>(
     ProviderCreate<T> definition, {
-    Qualifier qualifier,
+    Qualifier? qualifier,
     bool createdAtStart = false,
     bool override = false,
   }) {
@@ -50,7 +51,8 @@ extension BlocModuleExtension on Module {
         qualifier: qualifier,
         createdAtStart: createdAtStart,
         override: override);
-    beanDefinition.onClose((value) => value.dispose());
+    beanDefinition.onClose((value) => value.dispose(),
+        onDisposeUnitialized: () {});
     return beanDefinition;
   }
 }
@@ -71,14 +73,15 @@ extension ScopeSetBlocExtension on ScopeDSL {
   @deprecated
   ProviderDefinition<T> scopedBloc<T extends Disposable>(
     ProviderCreate<T> create, {
-    Qualifier qualifier,
+    Qualifier? qualifier,
     bool createdAtStart = false,
     bool override = false,
   }) {
     var providerDefinition =
         scoped<T>(create, qualifier: qualifier, override: override);
 
-    providerDefinition.onClose((bloc) => bloc.dispose());
+    providerDefinition.onClose((bloc) => bloc.dispose(),
+        onDisposeUnitialized: () {});
     return providerDefinition;
   }
 }
